fix: Correctly handle registry prefix
When listing by prefix, this should ensure the `/` is added at the end,
otherwise URLs for subdomains, etc are also included in the results,
leading to finding multiple logins, when in reality only one exists.

As an example, logging in to Docker Hub via `docker login` results in
the following entries  in the gopass store:

```
gopass
└── io.container.registries/
    ├── aHR0cHM6Ly9pbmRleC5kb2NrZXIuaW8vdjEv/
    │   └── dGVzdHVzZXIK
    ├── aHR0cHM6Ly9pbmRleC5kb2NrZXIuaW8vdjEvYWNjZXNzLXRva2Vu/
    │   └── dGVzdHVzZXIK
    └── aHR0cHM6Ly9pbmRleC5kb2NrZXIuaW8vdjEvcmVmcmVzaC10b2tlbg==/
        └── dGVzdHVzZXIK
```

Entries are created for:

https://index.docker.io/v1/
https://index.docker.io/v1/access-token
https://index.docker.io/v1/refresh-token%

When listing these are seen as multiple logins, where in reality they
are a single login (in the case above for `testuser`).